backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.2
python-dotenv==1.0.1
sqlalchemy==2.0.32
pymongo==4.7.3
psycopg[binary]==3.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart==0.0.12
pydantic[email]==2.9.2
requests==2.31.0
openai==1.52.2
httpx==0.27.2
psycopg2-binary==2.9.11